% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayeredMAplotfunctions2.R
\name{LayeredMAplot}
\alias{LayeredMAplot}
\title{グラフ描画用function}
\usage{
LayeredMAplot(MA_group_tbl, selected_vars, layer_val, ...)
}
\arguments{
\item{MA_group_tbl}{`make_grouped_MA_tbl`で生成された層化変数でgroup化した分析対象の含まれたdf}

\item{selected_vars}{分析対象のMAの変数名、もしくは変数番号}

\item{layer_val}{層化する変数名(char)}
}
\description{
層化MAplot（Dot plot）
}
