% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayeredMAplotfunctions2.R
\name{make_grouped_MA_tbl}
\alias{make_grouped_MA_tbl}
\title{Input}
\usage{
make_grouped_MA_tbl(df, selected_vars, layer_val, ...)
}
\arguments{
\item{df}{入力データフレーム（全体）df MAは、1/0にrecodeしてあること}

\item{selected_vars}{選択したMA変数（Q7#1...）chars 文字列でも変数番号でもよい}

\item{layer_val}{層化変数名 文字列 char

Output}
}
\value{
df format
}
\description{
Input
}
