% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayeredMAplotfunctions2.R
\name{plot_layered_facet_MA}
\alias{plot_layered_facet_MA}
\title{層化MAplot/facet の呼び出し方法}
\usage{
plot_layered_facet_MA(df, selected_vars, layer_val, sel = NULL, ...)
}
\arguments{
\item{df}{分析対象の含まれたdf}

\item{selected_vars}{分析対象のMAの変数名、もしくは変数番号}

\item{layer_val}{層化する変数名(char)}

\item{sel}{層化変数で利用するカテゴリ番号。defaultはすべて}
}
\value{
list(tbl = MAtbl, plot = p)
res <- plot_layered_MA(df, 50:70, "gender")
res$plot   # グラフ
res$tbl    # データ
}
\description{
層化MAplot/facet の呼び出し方法
}
