% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayeredMAplotfunctions2.R
\name{plot_layered_MA}
\alias{plot_layered_MA}
\title{Oct16,2025 版に対して、以下の改良
　層化変数を変数番号で呼び出せるようにした
　エラーバーを選択表示できるようにした。なし、標本誤差、標準偏差、標準誤差
　plotのlegendを選択できるようにした。ない、右、上、左、下
　
層化MAplotの呼び出し方法}
\usage{
plot_layered_MA(df, selected_vars, layer_var, sel = NULL, legend_pos = "right")
}
\arguments{
\item{df}{分析対象の含まれたdf}

\item{selected_vars}{分析対象のMAの変数名、もしくは変数番号}

\item{sel}{層化変数で利用するカテゴリ番号。defaultはすべて}

\item{legend_pos}{凡例の表示場所。"none","top","bottom",default は"right"}

\item{layer_val}{層化する変数名(char)　<- 変数番号もOKにする2025/10/20}
}
\value{
list(tbl = MAtbl, plot = p)
res <- plot_layered_MA(df, 50:70, "gender")　50:70は、サンプル。selected_varsで入力
res$plot   # グラフ
res$tbl    # データ
}
\description{
Oct16,2025 版に対して、以下の改良
　層化変数を変数番号で呼び出せるようにした
　エラーバーを選択表示できるようにした。なし、標本誤差、標準偏差、標準誤差
　plotのlegendを選択できるようにした。ない、右、上、左、下
　
層化MAplotの呼び出し方法
}
